Remove SLF4J and Lombok dependencies
Having SLF4J as a dependency causes the finished jar to more than double in file size. It is also unnecessary as Lahc should not log anything.
As a consequence of this, Logback was also removed from the test dependencies, as it will not be used.

Lombok was removed because it will not be necessary - code can be written, and saving yourself from having to write 1 simple line of code is not enough to justify an entire dependency.
Sure, it doesn't have the same problem as SLF4J with adding to final file size, however it is completely unnecessary. The only usage of it anywhere in the code was for tests, even though JUnit has its own logging system.